cleaning up Chapter 3
I'm still trying to understand the crossword puzzle tutorial. So I'm going through it again and testing the instructions.

One problem with the Readme was that it showed a command with a deposit, but this Chapter 3 contract's `new_puzzle` method doesn't expect a deposit because it's using:

```
// 5 Ⓝ in yoctoNEAR
const PRIZE_AMOUNT: u128 = 5_000_000_000_000_000_000_000_000;
```

Even when changing the `new_puzzle` command to not offer a deposit, I get this error:

> "Smart contract panicked: panicked at 'Failed to deserialize input from JSON.: Error(\"missing field `args`\", line: 1, column: 68)', src/lib.rs:133:1"}}

So more updates to this Readme (or the contract) are still required before this lesson actually works.